requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.5
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.4.0
Deprecated==1.2.14
dill==0.3.6
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.2.0
django-timezone-field==5.0
flake8==6.0.0
isort==5.12.0
kombu==5.3.0rc2
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.4
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
